requirements.txt
pandas>=2.3.0
numpy>=2.2.6
scikit-learn>=1.7.0
torch>=2.7.1
transformers>=4.35.2
sentencepiece>=0.2.0
pyarabic>=0.6.15
emoji>=2.14.1
PyStemmer>=3.0.0
nltk>=3.9.1
tqdm>=4.67.1
optuna>=4.3.0
